Use 64-bit video IDs

+++ src/dotnet/DataCache.cs
@@ -8,7 +8,7 @@ namespace StreamXRef
         public string ApiKey { get; set; }
         public Dictionary<String, Int32> UserInfoCache { get; set; }
         public Dictionary<String, ClipObject> ClipInfoCache { get; set; }
-        public Dictionary<Int32, DateTime> VideoInfoCache { get; set; }
+        public Dictionary<Int64, DateTime> VideoInfoCache { get; set; }
 
         public int GetTotalCount() => UserInfoCache.Count + ClipInfoCache.Count + VideoInfoCache.Count;
 
@@ -17,14 +17,14 @@ namespace StreamXRef
             ApiKey = "";
             UserInfoCache = new Dictionary<String, Int32>(StringComparer.OrdinalIgnoreCase);
             ClipInfoCache = new Dictionary<String, ClipObject>(StringComparer.OrdinalIgnoreCase);
-            VideoInfoCache = new Dictionary<Int32, DateTime>();
+            VideoInfoCache = new Dictionary<Int64, DateTime>();
         }
     }
 
     public class ClipObject
     {
         public int Offset { get; set; }
-        public int VideoID { get; set; }
+        public Int64 VideoID { get; set; }
         public DateTime Created { get; set; }
         public Dictionary<String, String> Mapping { get; set; }
